Improve compatibility for graceful shutdown.
