+++ dubbo-config/dubbo-config-api/src/main/java/com/alibaba/dubbo/config/ProtocolConfig.java
@@ -460,4 +460,13 @@ public class ProtocolConfig extends AbstractConfig {
         }
     }
 
+    /**
+     * This method will do nothing.
+     * It should be deleted in the next major version, say 2.7.x.
+     * Just for compatibility.
+     */
+    @Deprecated
+    public static void destroyAll() {
+        // DO NOTHING
+    }
 }
